kyros-praxis/packages/service-registry/requirements.txt
pytest==7.4.3
pytest-asyncio==0.23.8
pytest-cov==5.0.0
httpx==0.27.0
# testcontainers-postgres==1.0.0
fastapi>=0.104.1
pydantic>=2.5
sqlalchemy>=2.0
python-jose[cryptography]
passlib[bcrypt]
redis
uvicorn[standard]
